cleanup: move InternalFuture into the server module
