+++ backend/server/src/main/java/ai/verta/modeldb/experiment/FutureExperimentServiceImpl.java
@@ -10,8 +10,8 @@ import ai.verta.modeldb.common.exceptions.InvalidArgumentException;
 import ai.verta.modeldb.common.exceptions.ModelDBException;
 import ai.verta.modeldb.common.exceptions.NotFoundException;
 import ai.verta.modeldb.common.futures.FutureExecutor;
-import ai.verta.modeldb.common.futures.InternalFuture;
 import ai.verta.modeldb.project.FutureProjectDAO;
+import ai.verta.modeldb.utils.InternalFuture;
 import ai.verta.modeldb.utils.InternalFutureGrpc;
 import com.google.protobuf.Value;
 import com.google.rpc.Code;
