+++ backend/server/src/main/java/ai/verta/modeldb/experimentRun/subtypes/FilterPrivilegedVersionedInputsHandler.java
@@ -4,8 +4,8 @@ import ai.verta.common.ModelDBResourceEnum;
 import ai.verta.modeldb.VersioningEntry;
 import ai.verta.modeldb.common.futures.FutureExecutor;
 import ai.verta.modeldb.common.futures.FutureJdbi;
-import ai.verta.modeldb.common.futures.InternalFuture;
 import ai.verta.modeldb.interfaces.CheckEntityPermissionBasedOnResourceTypesFunction;
+import ai.verta.modeldb.utils.InternalFuture;
 import ai.verta.uac.ModelDBActionEnum;
 import java.util.ArrayList;
 import java.util.Collections;
