+++ backend/server/src/main/java/ai/verta/modeldb/experimentRun/subtypes/PredicatesHandler.java
@@ -13,8 +13,8 @@ import ai.verta.modeldb.common.exceptions.ModelDBException;
 import ai.verta.modeldb.common.exceptions.UnimplementedException;
 import ai.verta.modeldb.common.futures.Future;
 import ai.verta.modeldb.common.futures.FutureExecutor;
-import ai.verta.modeldb.common.futures.InternalFuture;
 import ai.verta.modeldb.common.query.QueryFilterContext;
+import ai.verta.modeldb.utils.InternalFuture;
 import com.google.protobuf.Value;
 import java.util.Arrays;
 import java.util.LinkedList;
