+++ backend/server/src/main/java/ai/verta/modeldb/experimentRun/subtypes/SortingHandler.java
@@ -2,10 +2,10 @@ package ai.verta.modeldb.experimentRun.subtypes;
 
 import ai.verta.modeldb.ModelDBConstants;
 import ai.verta.modeldb.common.exceptions.InvalidArgumentException;
-import ai.verta.modeldb.common.futures.InternalFuture;
 import ai.verta.modeldb.common.query.OrderColumn;
 import ai.verta.modeldb.common.query.OrderTable;
 import ai.verta.modeldb.common.query.QueryFilterContext;
+import ai.verta.modeldb.utils.InternalFuture;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
