+++ backend/server/src/main/java/ai/verta/modeldb/interfaces/CheckEntityPermissionBasedOnResourceTypesFunction.java
@@ -1,7 +1,7 @@
 package ai.verta.modeldb.interfaces;
 
 import ai.verta.common.ModelDBResourceEnum;
-import ai.verta.modeldb.common.futures.InternalFuture;
+import ai.verta.modeldb.utils.InternalFuture;
 import ai.verta.uac.ModelDBActionEnum;
 import java.util.List;
