+++ backend/server/src/main/java/ai/verta/modeldb/project/FutureProjectDAO.java
@@ -62,6 +62,7 @@ import ai.verta.modeldb.experimentRun.subtypes.PredicatesHandler;
 import ai.verta.modeldb.experimentRun.subtypes.SortingHandler;
 import ai.verta.modeldb.experimentRun.subtypes.TagsHandler;
 import ai.verta.modeldb.project.subtypes.CreateProjectHandler;
+import ai.verta.modeldb.utils.InternalFuture;
 import ai.verta.modeldb.utils.ModelDBUtils;
 import ai.verta.modeldb.utils.RdbmsUtils;
 import ai.verta.uac.Action;
